code optimization * move used readers from reproduction to io/dataset_loader.py (API shall not call anything from reproduction/)
